Remove unlocked items input and add a button on the card overlay to lock & unlock items
This also changed the way we store unlocked items in memory. Instead of
an array with all of the unlocked items directly in it we have an array
of ranges which more directly maps to how many items in the game are
unlocked. This should result in relatively few ranges (exactly 1 when
every item is eventually unlocked) so figuring out if a particular item
is unlocked should be a relatively fast operation. This is important
as it's a very common check when filtering items when certain filters
are selected and also helps on initial page load and when the URL hash
is updated directly